Crate spacetimedb_sats

Re-exports

pub use algebraic_type::AlgebraicType;

pub use algebraic_value::AlgebraicValue;

pub use algebraic_value::F32;

pub use algebraic_value::F64;

pub use array_type::ArrayType;

pub use array_value::ArrayValue;

pub use product_type::ProductType;

pub use product_type_element::ProductTypeElement;

pub use product_value::ProductValue;

pub use sum_type::SumType;

pub use sum_type_variant::SumTypeVariant;

pub use sum_value::SumValue;

pub use typespace::GroundSpacetimeType;

pub use typespace::SpacetimeType;

pub use typespace::Typespace;

pub use de::Deserialize;

pub use ser::Serialize;

Modules

algebraic_type

algebraic_value

array_type

array_value

bsatn

buffer

Minimal utility for reading & writing the types we need to internal storage, without relying on types in third party libraries like bytes::Bytes, etc. Meant to be kept slim and trim for use across both native and WASM.

convert

de

hash

hex

Allocation-free hex formatting.

layout

Defines Layout, which encompasses the fixed size and alignment of an object, e.g., a row, or a column, or some other sub-division of a row.

meta_type

Provides the MetaType trait.

primitives

product_type

product_type_element

product_value

satn

ser

size_of

sum_type

sum_type_variant

sum_value

time_duration

timestamp

typespace

Macros

impl_deserialize

Implements Deserialize for a type in a simplified manner.

impl_serialize

Implements Serialize for a type in a simplified manner.

impl_st

Implements SpacetimeType for a type in a simplified manner.

product

Constructs a product value from a list of fields with syntax product![v1, v2, ...].

Structs

AlgebraicTypeRef

A reference to an AlgebraicType within a Typespace.

IterWithTypespace

ValueWithType

A borrowed value combined with its type and typing context (Typespace).

WithTypespace

Adds a Typespace context atop of a borrowed type.

Traits

Value

The Value trait provides an abstract notion of a value.

Functions

hash_bsatn_array

Hashes every elem in the BSATN-encoded array value.

Type Aliases

i256

A 256-bit signed integer type.

u256

A 256-bit unsigned integer type.

Derive Macros

Deserialize

Serialize

SpacetimeType