spacetimedb_sats::algebraic_type::fmt

Function fmt_algebraic_type

pub fn fmt_algebraic_type(ty: &AlgebraicType) -> impl '_ + Display

Wraps the algebraic ty into a Displayable.

NOTE: You might ask: Why do we have a formatter and a notation for AlgebraicTypes if we don’t have an encoding for AlgebraicTypes?

This is because we just want an easier to read text format for algebraic types. This could just as easily take in an algebraic value, which represents an algebraic type and format it that way. It’s just more convenient to format it from the Rust type.