spacetimedb_sats::array_value

Enum ArrayValue

pub enum ArrayValue {

    Sum(Box<[SumValue]>),
    Product(Box<[ProductValue]>),
    Bool(Box<[bool]>),
    I8(Box<[i8]>),
    U8(Box<[u8]>),
    I16(Box<[i16]>),
    U16(Box<[u16]>),
    I32(Box<[i32]>),
    U32(Box<[u32]>),
    I64(Box<[i64]>),
    U64(Box<[u64]>),
    I128(Box<[i128]>),
    U128(Box<[u128]>),
    I256(Box<[i256]>),
    U256(Box<[u256]>),
    F32(Box<[F32]>),
    F64(Box<[F64]>),
    String(Box<[Box<str>]>),
    Array(Box<[ArrayValue]>),

}

An array value in “monomorphized form”.

Arrays are represented in this way monomorphized fashion for efficiency rather than unnecessary indirections and tags of AlgebraicValue. We can do this as we know statically that the type of each element is the same as arrays are homogenous dynamically sized product types.

Variants

Sum(Box<[SumValue]>)

An array of SumValues.

Product(Box<[ProductValue]>)

An array of ProductValues.

Bool(Box<[bool]>)

An array of bools.

I8(Box<[i8]>)

An array of i8s.

U8(Box<[u8]>)

An array of u8s.

I16(Box<[i16]>)

An array of i16s.

U16(Box<[u16]>)

An array of u16s.

I32(Box<[i32]>)

An array of i32s.

U32(Box<[u32]>)

An array of u32s.

I64(Box<[i64]>)

An array of i64s.

U64(Box<[u64]>)

An array of u64s.

I128(Box<[i128]>)

An array of i128s.

U128(Box<[u128]>)

An array of u128s.

I256(Box<[i256]>)

An array of i256s.

U256(Box<[u256]>)

An array of u256s.

F32(Box<[F32]>)

An array of totally ordered F32s.

F64(Box<[F64]>)

An array of totally ordered F64s.

String(Box<[Box<str>]>)

An array of UTF-8 strings.

Array(Box<[ArrayValue]>)

An array of arrays.

Implementations

impl ArrayValue

pub fn len(&self) -> usize

Returns the length of the array.

pub fn is_empty(&self) -> bool

Returns whether the array is empty.

pub fn iter_cloned(&self) -> ArrayValueIterCloned<'_>

Returns a cloning iterator on the elements of self as AlgebraicValues.