spacetimedb_sats

Module bsatn

Re-exports

pub use de::Deserializer;

pub use ser::Serializer;

pub use crate::buffer::DecodeError;

pub use ser::BsatnError as EncodeError;

Modules

de

eq

Defines the function eq_bsatn which equates lhs: &AlgebraicValue to rhs defined in BSATN.

ser

Traits

BufReservedFill

Provides a view over a buffer that can reserve an additional len bytes and then provide those as an uninitialized buffer to write into.

ToBsatn

Types that can be encoded to BSATN.

Functions

assert_is_primitive_type

Enforces that a type is a primitive. This is purely intended for use in crates/bindings-macro.

decode

Decode bytes to the value type of ty: S.

from_reader

Deserialize a T from the BSATN format in the buffered reader.

from_slice

Deserialize a T from the BSATN format in bytes.

to_len

Computes the size of val when BSATN encoding without actually encoding.

to_vec

Serialize value into a Vec<u8> in the BSATN format.

to_writer

Serialize value into the buffered writer w in the BSATN format.