spacetimedb_sats::bsatn

Trait BufReservedFill

pub trait BufReservedFill {
    // Required method
    unsafe fn reserve_and_fill(
        &mut self,
        len: usize,
        fill: impl FnOnce(&mut [MaybeUninit<u8>]),
    );
}

Provides a view over a buffer that can reserve an additional len bytes and then provide those as an uninitialized buffer to write into.

Required Methods

unsafe fn reserve_and_fill( &mut self, len: usize, fill: impl FnOnce(&mut [MaybeUninit<u8>]), )

Reserves space for len in self and then runs fill to fill it, adding len to the total length of self.

Safety

fill must initialize every byte in the slice.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.