spacetimedb_sats

Module de

Structs

BasicSmallVecVisitor

An implementation of [ArrayVisitor<'de, T>] where the output is a SmallVec<[T; N]>.

BasicVecVisitor

An implementation of [ArrayVisitor<'de, T>] where the output is a Vec<T>.

FDisplay

Turns a closure impl Fn(&mut Formatter) -> Result into a Displayable object.

NoneAccess

Deserializes none variant of an optional value.

Enums

ProductKind

What kind of product is this?

Traits

ArrayAccess

Provides an ArrayVisitor with access to each element of the array in the input.

ArrayVisitor

A visitor walking through a Deserializer for arrays.

Deserialize

A data structure that can be deserialized from any data format supported by the SpacetimeDB Algebraic Type System.

DeserializeOwned

A data structure that can be deserialized in SATS without borrowing any data from the deserializer.

DeserializeSeed

DeserializeSeed is the stateful form of the Deserialize trait.

Deserializer

A data format that can deserialize any data structure supported by SATS.

Error

The Error trait allows Deserialize implementations to create descriptive error messages belonging to the Deserializer against which they are currently running.

FieldNameVisitor

Visitor used to deserialize the name of a field.

GrowingVec

A vector with two operations: with_capacity and push.

NamedProductAccess

Provides a ProductVisitor with access to each element of the named product in the input.

ProductVisitor

A visitor walking through a Deserializer for products.

SeqProductAccess

Provides a ProductVisitor with access to each element of the unnamed product in the input.

SliceVisitor

A SliceVisitor is provided a slice T of some elements by a Deserializer and is tasked with translating this slice to the Output type.

SumAccess

Provides a SumVisitor access to the data of a sum in the input.

SumVisitor

A visitor walking through a Deserializer for sums.

VariantAccess

A visitor passed from SumAccess to SumVisitor::visit_sum which the latter uses to deserialize the data of a selected variant.

VariantVisitor

A visitor passed from SumVisitor to SumAccess::variant which the latter uses to decide what variant to deserialize.

Functions

array_visit

A basic implementation of ArrayVisitor::visit using the provided size hint.

fmt_fn

Turns a closure F: Fn(&mut Formatter) -> Result into a Displayable object.

Derive Macros

Deserialize