spacetimedb_sats::de

Trait DeserializeSeed

pub trait DeserializeSeed<'de> {
    type Output;

    // Required method
    fn deserialize<D: Deserializer<'de>>(
        self,
        deserializer: D,
    ) -> Result<Self::Output, D::Error>;
}

DeserializeSeed is the stateful form of the Deserialize trait.

Required Associated Types

type Output

The type produced by using this seed.

Required Methods

fn deserialize<D: Deserializer<'de>>( self, deserializer: D, ) -> Result<Self::Output, D::Error>

Equivalent to the more common Deserialize::deserialize associated function, except with some initial piece of data (the seed self) passed in.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.