spacetimedb_sats

Module layout

Defines Layout, which encompasses the fixed size and alignment of an object, e.g., a row, or a column, or some other sub-division of a row.

Layout annotated versions of SATS types are also provided, such as ProductTypeLayout and AlgebraicTypeLayout. These, and others, determine what the layout of objects typed at those types are. They also implement HasLayout which generalizes over layout annotated types.

Structs

Layout

The layout of a fixed object or the layout that fixed objects of a type will have.

ProductTypeElementLayout

A mirrior of ProductTypeElement annotated with a Layout.

ProductTypeLayout

A mirror of ProductType annotated with a Layout.

ProductTypeLayoutView

A mirror of ProductType annotated with a Layout.

RowTypeLayout

The type of a row, annotated with a Layout.

Size

The size of something in page storage in bytes.

SumTypeLayout

A mirrior of SumType annotated with a Layout.

SumTypeVariantLayout

A mirrior of SumTypeVariant annotated with a Layout.

Enums

AlgebraicTypeLayout

Mostly a mirror of AlgebraicType annotated with a Layout.

IncompatibleTypeLayoutError

Reason why two RowTypeLayouts are incompatible for an auto-migration.

IncompatibleTypeReason

PrimitiveType

Scalar types, i.e. bools, integers and floats. These types do not require a VarLenRef indirection when stored in a spacetimedb_table::table::Table.

VarLenType

Types requiring a VarLenRef indirection, i.e. strings, arrays, and maps.

Constants

MIN_ROW_ALIGN

Fixed-length row portions must also be sufficiently aligned to store a FreeCellRef.

MIN_ROW_SIZE

Fixed-length row portions must be at least large enough to store a FreeCellRef.

VAR_LEN_REF_LAYOUT

The layout of var-len objects. Aligned at a u16 which it has 2 of.

Traits

HasLayout

A type which knows what its layout is.

Functions

align_to

Aligns a base offset to the required_alignment (in the positive direction) and returns it.

row_size_for_bytes

Returns the minimum row size needed to store required_bytes accounting for the minimum row size and alignment.

row_size_for_type

Returns the minimum row size needed to store a T, accounting for the minimum row size and alignment.