spacetimedb_sats::layout

Enum AlgebraicTypeLayout

pub enum AlgebraicTypeLayout {
    Sum(SumTypeLayout),
    Product(ProductTypeLayout),
    Primitive(PrimitiveType),
    VarLen(VarLenType),
}

Mostly a mirror of AlgebraicType annotated with a Layout.

Notable differences from AlgebraicType:

Refs are not supported. Supporting recursive types remains a TODO(future-work). Note that the previous Spacetime datastore did not support recursive types in tables.

Scalar types (ty.is_scalar()) are separated into PrimitiveType (atomically-sized types like integers).

Variable length types are separated into VarLenType (strings, arrays, and maps). This separation allows cleaner pattern-matching, e.g. in HasLayout::layout, where VarLenType returns a static ref to VAR_LEN_REF_LAYOUT, and PrimitiveType dispatches on its variant to return a static ref to a type-specific Layout.

Variants

Sum(SumTypeLayout)

A sum type, annotated with its layout.

Product(ProductTypeLayout)

A product type, annotated with its layout.

Primitive(PrimitiveType)

A primitive type, annotated with its layout.

VarLen(VarLenType)

A variable length type, annotated with its layout.

Implementations

impl AlgebraicTypeLayout

pub fn is_sum(&self) -> bool

Returns true if this is a AlgebraicTypeLayout::Sum, otherwise false

pub fn as_sum_mut(&mut self) -> Option<&mut SumTypeLayout>

Optionally returns mutable references to the inner fields if this is a AlgebraicTypeLayout::Sum, otherwise None

pub fn as_sum(&self) -> Option<&SumTypeLayout>

Optionally returns references to the inner fields if this is a AlgebraicTypeLayout::Sum, otherwise None

pub fn into_sum(self) -> Result<SumTypeLayout, Self>

Returns the inner fields if this is a AlgebraicTypeLayout::Sum, otherwise returns back the enum in the Err case of the result

pub fn is_product(&self) -> bool

Returns true if this is a AlgebraicTypeLayout::Product, otherwise false

pub fn as_product_mut(&mut self) -> Option<&mut ProductTypeLayout>

Optionally returns mutable references to the inner fields if this is a AlgebraicTypeLayout::Product, otherwise None

pub fn as_product(&self) -> Option<&ProductTypeLayout>

Optionally returns references to the inner fields if this is a AlgebraicTypeLayout::Product, otherwise None

pub fn into_product(self) -> Result<ProductTypeLayout, Self>

Returns the inner fields if this is a AlgebraicTypeLayout::Product, otherwise returns back the enum in the Err case of the result

pub fn is_primitive(&self) -> bool

Returns true if this is a AlgebraicTypeLayout::Primitive, otherwise false

pub fn as_primitive_mut(&mut self) -> Option<&mut PrimitiveType>

Optionally returns mutable references to the inner fields if this is a AlgebraicTypeLayout::Primitive, otherwise None

pub fn as_primitive(&self) -> Option<&PrimitiveType>

Optionally returns references to the inner fields if this is a AlgebraicTypeLayout::Primitive, otherwise None

pub fn into_primitive(self) -> Result<PrimitiveType, Self>

Returns the inner fields if this is a AlgebraicTypeLayout::Primitive, otherwise returns back the enum in the Err case of the result

pub fn is_var_len(&self) -> bool

Returns true if this is a AlgebraicTypeLayout::VarLen, otherwise false

pub fn as_var_len_mut(&mut self) -> Option<&mut VarLenType>

Optionally returns mutable references to the inner fields if this is a AlgebraicTypeLayout::VarLen, otherwise None

pub fn as_var_len(&self) -> Option<&VarLenType>

Optionally returns references to the inner fields if this is a AlgebraicTypeLayout::VarLen, otherwise None

pub fn into_var_len(self) -> Result<VarLenType, Self>

Returns the inner fields if this is a AlgebraicTypeLayout::VarLen, otherwise returns back the enum in the Err case of the result

impl AlgebraicTypeLayout

pub const Bool: Self

pub const I8: Self

pub const U8: Self

pub const I16: Self

pub const U16: Self

pub const I32: Self

pub const U32: Self

pub const I64: Self

pub const U64: Self

pub const I128: Self

pub const U128: Self

pub const I256: Self

pub const U256: Self

pub const F32: Self

pub const F64: Self

pub const String: Self

impl AlgebraicTypeLayout

pub fn algebraic_type(&self) -> AlgebraicType

Convert an AlgebraicTypeLayout back into an AlgebraicType, removing layout information.

This operation is O(n) in the number of nodes in the argument, and may heap-allocate. It is intended for use in error paths, where performance is a secondary concern.