spacetimedb_sats::meta_type

Trait MetaType

pub trait MetaType {
    // Required method
    fn meta_type() -> AlgebraicType;
}

Rust types which represent components of the SATS type system and can themselves be represented as algebraic objects will implement MetaType.

A type’s meta-type is an AlgebraicType which can store the data associated with a definition of that type.

For example, the MetaType of ProductType is

AlgebraicType::product([(
    "elements",
    AlgebraicType::array(ProductTypeElement::meta_type()),
)])

Required Methods

fn meta_type() -> AlgebraicType

Returns the type structure of this type as an AlgebraicType.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.