spacetimedb_sats

Module satn

Structs

PsqlChars

PsqlType

A wrapper that remember the header of the tuple/struct and the current field

PsqlWrapper

A wrapper around a T: Satn providing Display and Debug implementations that uses postgres SATN formatting for T.

SatnError

An error occurred during serialization to the SATS data format.

SqlFormatter

An implementation of Serializer for SQL output.

TypedArrayFormatter

A formatter for arrays that uses the TypedWriter trait to write elements.

TypedNamedProductFormatter

A formatter for named products that uses the TypedWriter trait to write elements.

TypedSeqFormatter

A formatter for sequences that uses the TypedWriter trait to write elements.

TypedSerializer

A serializer that uses the TypedWriter trait to serialize values

Wrapper

A wrapper around a T: Satn providing Display and Debug implementations that uses the SATN formatting for T.

Enums

PsqlClient

Which client is used to format the SQL output?

PsqlPrintFmt

How format of the SQL output?

Traits

Satn

An extension trait for Serialize providing formatting methods.

TypedWriter

A trait for writing values, after the special types has been determined.