spacetimedb_sats::satn

Trait TypedWriter

pub trait TypedWriter {
    type Error: Error;

    // Required methods
    fn write<W: Display>(&mut self, value: W) -> Result<(), Self::Error>;
    fn write_bool(&mut self, value: bool) -> Result<(), Self::Error>;
    fn write_string(&mut self, value: &str) -> Result<(), Self::Error>;
    fn write_bytes(&mut self, value: &[u8]) -> Result<(), Self::Error>;
    fn write_hex(&mut self, value: &[u8]) -> Result<(), Self::Error>;
    fn write_timestamp(&mut self, value: Timestamp) -> Result<(), Self::Error>;
    fn write_duration(&mut self, value: TimeDuration) -> Result<(), Self::Error>;
    fn write_uuid(&mut self, value: Uuid) -> Result<(), Self::Error>;
    fn write_record(
        &mut self,
        fields: Vec<(Cow<'_, str>, PsqlType<'_>, ValueWithType<'_, AlgebraicValue>)>,
    ) -> Result<(), Self::Error>;
    fn write_variant(
        &mut self,
        tag: u8,
        ty: PsqlType<'_>,
        name: Option<&str>,
        value: ValueWithType<'_, AlgebraicValue>,
    ) -> Result<(), Self::Error>;

    // Provided method
    fn write_alt_record(
        &mut self,
        _ty: &PsqlType<'_>,
        _value: &ValueWithType<'_, ProductValue>,
    ) -> Result<bool, Self::Error> { ... }
}

A trait for writing values, after the special types has been determined.

This is used to write values that could have different representations depending on the output format, as defined by PsqlClient and PsqlPrintFmt.

Required Associated Types

type Error: Error

Required Methods

fn write<W: Display>(&mut self, value: W) -> Result<(), Self::Error>

Writes a value using ser::Serializer

fn write_bool(&mut self, value: bool) -> Result<(), Self::Error>

fn write_string(&mut self, value: &str) -> Result<(), Self::Error>

fn write_bytes(&mut self, value: &[u8]) -> Result<(), Self::Error>

fn write_hex(&mut self, value: &[u8]) -> Result<(), Self::Error>

fn write_timestamp(&mut self, value: Timestamp) -> Result<(), Self::Error>

fn write_duration(&mut self, value: TimeDuration) -> Result<(), Self::Error>

fn write_uuid(&mut self, value: Uuid) -> Result<(), Self::Error>

fn write_record( &mut self, fields: Vec<(Cow<'_, str>, PsqlType<'_>, ValueWithType<'_, AlgebraicValue>)>, ) -> Result<(), Self::Error>

fn write_variant( &mut self, tag: u8, ty: PsqlType<'_>, name: Option<&str>, value: ValueWithType<'_, AlgebraicValue>, ) -> Result<(), Self::Error>

Provided Methods

fn write_alt_record( &mut self, _ty: &PsqlType<'_>, _value: &ValueWithType<'_, ProductValue>, ) -> Result<bool, Self::Error>

Writes a value as an alternative record format, e.g., for use JSON inside SQL.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.