spacetimedb_sats

Module ser

Structs

ForwardNamedToSeqProduct

Forwards the implementation of a named product value to the implementation of the unnamed kind, thereby ignoring any field names.

Impossible

A type usable in one of the associated types of Serializer when the data format does not support the data.

Traits

Error

The base trait serialization error types must implement.

Serialize

A data structure that can be serialized into any data format supported by the SpacetimeDB Algebraic Type System.

SerializeArray

Returned from Serializer::serialize_array.

SerializeNamedProduct

Returned from Serializer::serialize_named_product.

SerializeSeqProduct

Returned from Serializer::serialize_seq_product.

Serializer

A data format that can deserialize any data structure supported by SATs.

Derive Macros

Serialize