spacetimedb_sats::ser

Trait SerializeSeqProduct

pub trait SerializeSeqProduct {
    type Ok;
    type Error: Error;

    // Required methods
    fn serialize_element<T: Serialize + ?Sized>(
        &mut self,
        element: &T,
    ) -> Result<(), Self::Error>;
    fn end(self) -> Result<Self::Ok, Self::Error>;
}

Returned from Serializer::serialize_seq_product.

This provides a continuation of sorts where you can call serialize_element however many times and then finally the end is reached.

Required Associated Types

type Ok

Must match the Ok type of any Serializer that uses this type.

type Error: Error

Must match the Error type of any Serializer that uses this type.

Required Methods

fn serialize_element<T: Serialize + ?Sized>( &mut self, element: &T, ) -> Result<(), Self::Error>

Serialize an unnamed product element.

fn end(self) -> Result<Self::Ok, Self::Error>

Consumes and finalizes the product serializer returning the Self::Ok data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.