spacetimedb_sats

Struct WithTypespace

pub struct WithTypespace<'a, T: ?Sized> { /* private fields */ }

Adds a Typespace context atop of a borrowed type.

Implementations

impl<'a> WithTypespace<'a, ProductType>

pub fn elements(&self) -> ElementsWithTypespace<'a>

pub fn with_values<I: IntoIterator<Item = &'a AlgebraicValue>>( &self, vals: I, ) -> ElementValuesWithType<'a, I::IntoIter>
where I::IntoIter: ExactSizeIterator,

impl<'a> WithTypespace<'a, ProductTypeElement>

pub fn algebraic_type(&self) -> WithTypespace<'a, AlgebraicType>

impl<T: ResolveRefs> WithTypespace<'_, T>

pub fn resolve_refs(self) -> Result<T::Output, TypeRefError>

impl<'a, T: ?Sized> WithTypespace<'a, T>

pub const fn new(typespace: &'a Typespace, ty: &'a T) -> Self

Wraps ty in a context combined with the typespace.

pub const fn empty(ty: &'a T) -> Self

Wraps ty in an empty context.

pub const fn ty(&self) -> &'a T

Returns the object that the context was created with.

pub const fn typespace(&self) -> &'a Typespace

Returns the typespace context.

pub fn with<'b, U>(&self, ty: &'b U) -> WithTypespace<'b, U>
where 'a: 'b,

Reuses the typespace we already have and returns ty: U wrapped with it.

pub fn with_value<'b, V: Value<Type = T>>( &self, val: &'b V, ) -> ValueWithType<'b, V>
where 'a: 'b,

Wraps val with the type and typespace context in self.

pub fn resolve(&self, r: AlgebraicTypeRef) -> WithTypespace<'a, AlgebraicType>

Returns the AlgebraicType that r resolves to in the context of our Typespace.

Panics if r is not known by our Typespace.

pub fn map<U: ?Sized>( &self, f: impl FnOnce(&'a T) -> &'a U, ) -> WithTypespace<'a, U>

Maps the object we’ve wrapped from &T -> &U in our context.

This can be used to e.g., project fields and through a structure. This provides an implementation of functor mapping for WithTypespace.