spacetimedb_sats

Trait Value

pub trait Value {
    type Type;
}

The Value trait provides an abstract notion of a value.

All we know about values abstractly is that they have a Type.

Required Associated Types

type Type

The type of this value.