spacetimedb_sats::typespace

Trait GroundSpacetimeType

pub trait GroundSpacetimeType {
    // Required method
    fn get_type() -> AlgebraicType;
}

A trait for Rust types that can be represented as an AlgebraicType with an empty typing context.

The returned AlgebraicType must have no free variables, that is, no AlgebraicTypeRefs in its tree at all.

Required Methods

fn get_type() -> AlgebraicType

Returns the AlgebraicType representation of Self.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.