spacetimedb_sats::typespace

Trait TypespaceBuilder

pub trait TypespaceBuilder {
    // Required method
    fn add(
        &mut self,
        typeid: TypeId,
        name: Option<&'static str>,
        make_ty: impl FnOnce(&mut Self) -> AlgebraicType,
    ) -> AlgebraicType;

    // Provided method
    fn add_type<T: SpacetimeType>(&mut self) -> AlgebraicType
       where Self: Sized { ... }
}

A trait for types that can build a Typespace.

Required Methods

fn add( &mut self, typeid: TypeId, name: Option<&'static str>, make_ty: impl FnOnce(&mut Self) -> AlgebraicType, ) -> AlgebraicType

Returns and adds a representation of type T: 'static as an AlgebraicType with an optional name to the typing context in self.

Provided Methods

fn add_type<T: SpacetimeType>(&mut self) -> AlgebraicType
where Self: Sized,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.