spacetimedb_sats::de

Trait Deserialize

pub trait Deserialize<'de>: Sized {
    // Required method
    fn deserialize<D: Deserializer<'de>>(
        deserializer: D,
    ) -> Result<Self, D::Error>;
}

A data structure that can be deserialized from any data format supported by the SpacetimeDB Algebraic Type System.

In most cases, implementations of Deserialize may be #[derive(Deserialize)]d.

The Deserialize trait in SATS performs the same function as serde::Deserialize in serde. See the documentation of serde::Deserialize for more information of the data model.

The lifetime 'de allows us to deserialize lifetime-generic types in a zero-copy fashion.

Do not manually implement this trait unless you know what you are doing. Implementations must be consistent with Serialize for T, SpacetimeType for T and Serialize, Deserialize for AlgebraicValue. Implementations that are inconsistent across these traits may result in data loss.

Required Methods

fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error>

Deserialize this value from the given deserializer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.