spacetimedb_sats::de

Trait VariantAccess

pub trait VariantAccess<'de>: Sized {
    type Error: Error;

    // Required method
    fn deserialize_seed<T: DeserializeSeed<'de>>(
        self,
        seed: T,
    ) -> Result<T::Output, Self::Error>;

    // Provided method
    fn deserialize<T: Deserialize<'de>>(self) -> Result<T, Self::Error> { ... }
}

A visitor passed from SumAccess to SumVisitor::visit_sum which the latter uses to deserialize the data of a selected variant.

Required Associated Types

type Error: Error

Required Methods

fn deserialize_seed<T: DeserializeSeed<'de>>( self, seed: T, ) -> Result<T::Output, Self::Error>

Called when deserializing the contents of a sum variant, and provided with a seed value.

Provided Methods

fn deserialize<T: Deserialize<'de>>(self) -> Result<T, Self::Error>

Called when deserializing the contents of a sum variant.

This method exists as a convenience for Deserialize implementations.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.